svg/text/parsing/text-anchor-invalid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Text: parsing text-anchor with invalid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/text.html#TextAnchorProperty" rel="help"/><h:meta content="text-anchor supports only the grammar 'start | middle | end'." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_invalid_value("text-anchor", "auto");
test_invalid_value("text-anchor", "start middle");

  ]]></script></svg>
